requirements.txt
discord.py
python-dotenv
spotipy
PyNaCl
yt-dlp